+++ CHANGELOG.md
@@ -10,7 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - model for D-Link cisco like CLI (@mirackle-spb)
 - model for Ruijie Networks RGOS devices (@spike77453)
 - pagination for http source (@davama)
-
+- model for CeraOS (Ceragon OS)
 + Added ability to send mail with the Docker container
 + Documentation to send mail with hooks